Refactor public talks index component
