+++ app/components/public_talks/talks/index_page_component.rb
@@ -7,7 +7,6 @@ class PublicTalks::Talks::IndexPageComponent < PageComponent
     @sql = talks.to_sql # for testing purposes
     @talks = talks
     setup_breadcrumb
-    @week = MeetingWeek.new
   end
 
   def showing_talks_since
@@ -18,43 +17,6 @@ class PublicTalks::Talks::IndexPageComponent < PageComponent
 
   def new_link
     link_to(t('app.links.new'), new_public_talks_talk_path)
-  end
-
-  def edit_link(talk)
-    link_to(t('app.links.edit'), edit_public_talks_talk_path(talk))
-  end
-
-  def destroy_link(talk)
-    link_to(
-      t('app.links.delete'), public_talks_talk_path(talk),
-      data: { method: :delete, confirm: delete_confirmation(talk) }
-    )
-  end
-
-  def show_link(talk)
-    link_to(t('app.links.view'), public_talks_talk_path(talk))
-  end
-
-  def speaker_name(talk)
-    "#{talk&.speaker&.name} (#{talk&.speaker&.congregation&.name})"
-  end
-
-  def theme(talk)
-    talk.theme_object
-  end
-
-  def classes(talk)
-    classes = []
-
-    if talk.congregation.present?
-      classes << talk.congregation.local? ? 'local' : 'outcoing'
-    end
-
-    if @week.cover?(talk.date)
-      classes << 'current'
-    end
-
-    classes.join(' ')
   end
 
   private
@@ -69,10 +31,6 @@ class PublicTalks::Talks::IndexPageComponent < PageComponent
     rescue StandardError
       nil
     end
-  end
-
-  def delete_confirmation(talk)
-    t('app.messages.confirm_delete_x', record: talk.summary)
   end
 
   def setup_breadcrumb
