+++ app/components/public_talks/talks/list_item_component.html.erb
@@ -1,0 +1,33 @@
+<div class="item <%= classes(talk) %>">
+  <div class="row">
+    <div class="col-sm-12 col-lg-12 title mb-2">
+      <%= theme(talk) %>
+      <%= render DropdownMenuComponent.new do |component| %>
+        <% component.item do %>
+          <%= show_link %>
+        <% end %>
+        <% component.item do %>
+          <%= edit_link %>
+        <% end %>
+        <% component.item do %>
+          <%= destroy_link %>
+        <% end %>
+      <% end %>
+    </div>
+  </div>
+  <div class="row small">
+    <div class="col-sm-12 col-lg-4">
+      <% if talk.speaker.present? %>
+        <%= icon('person') { speaker_name(talk) } %>
+      <% end %>
+    </div>
+    <div class="col-sm-12 col-lg-4">
+      <% if talk.congregation.present? %>
+        <%= icon('shop') { talk.congregation.name } %>
+      <% end %>
+    </div>
+    <div class="col-sm-12 col-lg-4">
+      <%= icon('calendar-date') { l(talk.date, format: :short) } %>
+    </div>
+  </div>
+</div>
